Create bhum1234
Bhumij language is considered as a distinct language and is not a dialect of Mundari. The Government of India has recognised it as a distinct language, and published Bhumij Primer and Mundari Primer separately for primary education in Bhumij language. It is also recognised as official second state language of Jharkhand, India,